+++ cli/changelog.md
@@ -1,7 +1,24 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
 ## [Unreleased]
-- None.
+
+## [0.1.1] - 2023-02-17
+### Fixed
+- Changes were not merged correctly.
+
+## [0.1.0] - 2023-02-17
+### Added
+- Add name-prefix and ident-prefix from command line. 
+- Add replacement values in templates.
+- Add text template output for files.
+### Fixed
+- Fix whitespace handling in text output.
 
 ## [0.0.3] - 2022-12-05
 ### Added